+++ src/sender.cpp
@@ -1,13 +1,12 @@
-#include <iostream>
-#include <thread>
 #include "rtc/rtc.hpp"
+#include <iostream>
 #include <nlohmann/json.hpp>
+#include <thread>
 
-#include "sender.hpp"
 #include "sdp_file.hpp"
+#include "sender.hpp"
 
-Sender::Sender()
-{
+Sender::Sender() {
     std::remove(answer_path_.c_str());
     std::remove(offer_path_.c_str());
 
@@ -16,14 +15,12 @@ Sender::Sender()
     reg_callbacks();
 }
 
-void Sender::reg_callbacks()
-{
-    pc_->onStateChange([](rtc::PeerConnection::State state) {
-        std::cout << "ICE state: " << state  << std::endl;
-    });
+void Sender::reg_callbacks() {
+    pc_->onStateChange(
+        [](rtc::PeerConnection::State state) { std::cout << "ICE state: " << state << std::endl; });
 
     pc_->onGatheringStateChange([this](rtc::PeerConnection::GatheringState state) {
-        std::cout << "Gathering state: " << state  << std::endl;
+        std::cout << "Gathering state: " << state << std::endl;
         if (state == rtc::PeerConnection::GatheringState::Complete) {
             write_file(offer_path_, std::string(pc_->localDescription().value()));
             std::cout << "Offer written to: " << offer_path_ << std::endl;
@@ -32,31 +29,27 @@ void Sender::reg_callbacks()
 
     try {
         dc_ = pc_->createDataChannel("telemetry");
-    } catch (const std::exception& e) {
+    } catch (const std::exception &e) {
         std::cout << e.what() << std::endl;
         pc_->close();
         return;
     }
 
-    dc_->onOpen([this]() {
-        std::cout << "DataChannel opened: " << dc_->label() << std::endl;
-    });
+    dc_->onOpen([this]() { std::cout << "DataChannel opened: " << dc_->label() << std::endl; });
 
-    dc_->onClosed([this]() {
-        std::cout << "DataChannel closed: " << dc_->label() << std::endl;
-    });
+    dc_->onClosed([this]() { std::cout << "DataChannel closed: " << dc_->label() << std::endl; });
 }
 
-void Sender::send_message()
-{
+void Sender::send_message() {
     while (!dc_->isOpen())
-    std::this_thread::sleep_for(std::chrono::milliseconds(100));
+        std::this_thread::sleep_for(std::chrono::milliseconds(100));
 
     static uint64_t seq = 0;
     while (dc_->isOpen()) {
-        auto now      = std::chrono::system_clock::now().time_since_epoch();
+        auto now = std::chrono::system_clock::now().time_since_epoch();
         uint64_t secs = std::chrono::duration_cast<std::chrono::seconds>(now).count();
-        uint32_t ns   = std::chrono::duration_cast<std::chrono::nanoseconds>(now).count() % 1'000'000'000;
+        uint32_t ns = static_cast<uint32_t>(
+            std::chrono::duration_cast<std::chrono::nanoseconds>(now).count() % 1'000'000'000);
 
         Message msg{++seq, {secs, ns}, "hello"};
         dc_->send(nlohmann::json(msg).dump());
@@ -65,13 +58,12 @@ void Sender::send_message()
     }
 }
 
-void Sender::start()
-{
+void Sender::start() {
     std::cout << "Waiting for " << answer_path_ << std::endl;
     auto answer_sdp = read_file(answer_path_);
     auto deadline = std::chrono::steady_clock::now() + std::chrono::minutes(1);
 
-    while (!answer_sdp){
+    while (!answer_sdp) {
         if (std::chrono::steady_clock::now() > deadline) {
             std::cout << "Timeout waiting for answer" << std::endl;
             pc_->close();
@@ -82,9 +74,9 @@ void Sender::start()
     }
 
     try {
-        pc_->setRemoteDescription(rtc::Description(answer_sdp.value(),
-                                                   rtc::Description::Type::Answer));
-    } catch (const std::exception& e) {
+        pc_->setRemoteDescription(
+            rtc::Description(answer_sdp.value(), rtc::Description::Type::Answer));
+    } catch (const std::exception &e) {
         std::cout << "Faild to get SDP: " << e.what() << std::endl;
         pc_->close();
         return;
@@ -93,8 +85,7 @@ void Sender::start()
     send_message();
 }
 
-void Sender::stop()
-{
+void Sender::stop() {
     while (pc_->state() != rtc::PeerConnection::State::Closed &&
            pc_->state() != rtc::PeerConnection::State::Failed &&
            pc_->state() != rtc::PeerConnection::State::Disconnected) {
@@ -102,13 +93,9 @@ void Sender::stop()
     }
 }
 
-Sender::~Sender()
-{
-    stop();
-}
+Sender::~Sender() { stop(); }
 
-int main()
-{
+int main() {
     rtc::InitLogger(rtc::LogLevel::Warning);
     Sender sender;
     sender.start();
